completed category and publisher controller

+++ src/category/category.service.ts
@@ -1,14 +1,14 @@
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'prisma/prisma.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
-import { UpdateCategoryDto } from './dto/update-cateogory.dto';
+import { UpdateCategoryDto } from './dto/update-category.dto';
 
 @Injectable()
 export class CategoryService {
   constructor(private prisma: PrismaService) {}
 
-  createCatory(dto: CreateCategoryDto) {
+  createCategory(dto: CreateCategoryDto) {
     return this.prisma.category.create({
       data: {
         ...dto,
@@ -24,6 +24,16 @@ export class CategoryService {
     categoryId: number,
     updateCategoryDto: UpdateCategoryDto,
   ) {
+    //Get category by Id
+    const category = await this.prisma.category.findUnique({
+      where: {
+        id: categoryId,
+      },
+    });
+
+    if (!category || category.id !== categoryId)
+      throw new ForbiddenException('Category not found');
+
     return await this.prisma.category.update({
       where: {
         id: categoryId,
@@ -35,6 +45,16 @@ export class CategoryService {
   }
 
   async deleteCategory(categoryId: number) {
+    //Get category by Id
+    const category = await this.prisma.category.findUnique({
+      where: {
+        id: categoryId,
+      },
+    });
+
+    if (!category || category.id !== categoryId)
+      throw new ForbiddenException('Category not found');
+
     return this.prisma.category.delete({
       where: {
         id: categoryId,
